MCPU without Jacobian
This is MCPU as it stands on April 9 2019, in which we have a grid of replica exchanges, and everything works as I have been doing over the last few months, and I have also modified define.h to set:

#define CLUSTER_MOVE		0.0  #define USE_CLUSTER		0

In effect turning off the function loopmove and knowledge based moves. This is compiled in Fold potential MPI (there is also a version, fold_potential_mpi2, in whcih the loopmovefunction is called with probability CLUSTERMOVE = 0.3333, but USE_CLUSTER is set to 0 so you never acutally do a knowledge based move, but still take larger steps).

The thing that isn't here is the correct acceptance criterion for local moves, whcih should incorporate the Jacobian factor and number of loop closure solutions